Deal with case when no schemas found (no services online)
