+++ gateway/src/schema-stitching/errors.ts
@@ -1,0 +1,6 @@
+export class NoRemoteSchemasFound extends Error {
+    constructor(message) {
+        super(message);
+        this.name = "NoRemoteSchemasFound";
+    }
+}
